Extract ClanBattle and GuildExpedition request handlers
